requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
sympy>=1.12

